Minor code cleanup and add versions

+++ common/src/main/java/foundry/veil/api/flare/data/model/ShellElementRotation.java
@@ -6,11 +6,17 @@ import foundry.veil.api.util.CodecUtil;
 import net.minecraft.core.Direction;
 import org.joml.Vector3fc;
 
-public record ShellElementRotation(Vector3fc origin, Direction.Axis axis, int angle) {
+/**
+ * @param origin The origin of the rotation
+ * @param axis   The axis to rotate about
+ * @param angle  The angle to apply in degrees
+ * @since 2.5.0
+ */
+public record ShellElementRotation(Vector3fc origin, Direction.Axis axis, float angle) {
 
     public static final Codec<ShellElementRotation> CODEC = RecordCodecBuilder.create(instance -> instance.group(
             CodecUtil.VECTOR3FC_CODEC.fieldOf("origin").forGetter(ShellElementRotation::origin),
             Direction.Axis.CODEC.fieldOf("axis").forGetter(ShellElementRotation::axis),
-            Codec.INT.fieldOf("angle").forGetter(ShellElementRotation::angle)
+            Codec.FLOAT.fieldOf("angle").forGetter(ShellElementRotation::angle)
     ).apply(instance, ShellElementRotation::new));
 }
